--- VRTRAKILL.Plugin/Plugin/Patches/HelpersP.cs
@@ -1,17 +1,0 @@
-﻿using HarmonyLib;
-
-namespace VRBasePlugin.Patches
-{
-    [HarmonyPatch] internal class HelpersP
-    {
-        [HarmonyPrefix]
-        [HarmonyPatch(typeof(CameraController), nameof(CameraController.Update))]
-        [HarmonyPatch(typeof(CameraFrustumTargeter), nameof(CameraFrustumTargeter.Update))]
-        [HarmonyPatch(typeof(CameraFrustumTargeter), nameof(CameraFrustumTargeter.LateUpdate))]
-        static bool DoNothing()
-        {
-            // do nothing
-            return false;
-        }
-    }
-}
